backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
redis==5.1.1
celery==5.4.0
websockets==14.1
pandas==2.2.3
python-dateutil==2.9.0.post0
stripe==11.4.0
email-validator==2.2.0
google-auth==2.29.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httpx==0.27.0

